+++ packages/abi-typegen/src/abi/functions/Function.ts
@@ -2,6 +2,7 @@ import type { IFunction, JsonAbiFunction, IFunctionAttributes } from '../../inde
 import { TargetEnum } from '../../types/enums/TargetEnum';
 import type { IType } from '../../types/interfaces/IType';
 import { findType } from '../../utils/findType';
+import { resolveInputLabel } from '../../utils/getTypeDeclaration';
 import { parseTypeArguments } from '../../utils/parseTypeArguments';
 import { EmptyType } from '../types/EmptyType';
 
@@ -35,22 +36,7 @@ export class Function implements IFunction {
       .map((input) => {
         const { name, type: typeId, typeArguments } = input;
 
-        const type = findType({ types, typeId });
-
-        let typeDecl: string;
-
-        if (typeArguments) {
-          // recursively process child `typeArguments`
-          typeDecl = parseTypeArguments({
-            types,
-            target: TargetEnum.INPUT,
-            parentTypeId: typeId,
-            typeArguments,
-          });
-        } else {
-          // or just collect type declaration
-          typeDecl = type.attributes.inputLabel;
-        }
+        const typeDecl = resolveInputLabel(types, typeId, typeArguments);
 
         // assemble it in `[key: string]: <Type>` fashion
         if (shouldPrefixParams) {
