+++ packages/abi-typegen/src/templates/script/factory.ts
@@ -2,7 +2,6 @@ import { ErrorCode, FuelError } from '@fuel-ts/errors';
 
 import type { Abi } from '../../abi/Abi';
 import { renderHbsTemplate } from '../renderHbsTemplate';
-import { formatConfigurables } from '../utils/formatConfigurables';
 import { formatEnums } from '../utils/formatEnums';
 import { formatImports } from '../utils/formatImports';
 import { formatStructs } from '../utils/formatStructs';
@@ -32,7 +31,6 @@ export function renderFactoryTemplate(params: { abi: Abi }) {
   const { enums } = formatEnums({ types });
   const { structs } = formatStructs({ types });
   const { imports } = formatImports({ types, baseMembers: ['Script', 'Account'] });
-  const { formattedConfigurables } = formatConfigurables({ configurables });
 
   const { prefixedInputs: inputs, output } = func.attributes;
 
@@ -47,7 +45,7 @@ export function renderFactoryTemplate(params: { abi: Abi }) {
       hexlifiedBinString,
       capitalizedName,
       imports,
-      formattedConfigurables,
+      configurables,
       commonTypesInUse: commonTypesInUse.join(', '),
     },
   });
