+++ packages/abi-typegen/src/types/interfaces/IConfigurable.ts
@@ -1,8 +1,4 @@
-import type { IType } from './IType';
-import type { JsonAbiConfigurable } from './JsonAbi';
-
 export interface IConfigurable {
   name: string;
-  type: IType;
-  rawAbiConfigurable: JsonAbiConfigurable;
+  inputLabel: string;
 }
